requirements.txt
absl-py==0.13.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.0.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mediapipe==0.8.7
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
pip==21.0.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.6.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.1.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
Send2Trash==1.8.0
setuptools==52.0.0.post20210125
six==1.15.0
smmap==4.0.0
streamlit==0.87.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wheel==0.37.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
